+++ src/main/kotlin/app/revanced/patches/music/premium/backgroundplay/patch/BackgroundPlayPatch.kt
@@ -9,18 +9,16 @@ import app.revanced.patcher.patch.BytecodePatch
 import app.revanced.patcher.patch.PatchResult
 import app.revanced.patcher.patch.PatchResultSuccess
 import app.revanced.patcher.patch.annotations.Patch
-import app.revanced.patches.music.premium.backgroundplay.annotations.BackgroundPlayCompatibility
+import app.revanced.patches.music.annotations.MusicCompatibility
 import app.revanced.patches.music.premium.backgroundplay.fingerprints.BackgroundPlaybackDisableFingerprint
 
 @Patch
 @Name("background-play")
 @Description("Enables playing music in the background.")
-@BackgroundPlayCompatibility
+@MusicCompatibility
 @Version("0.0.1")
 class BackgroundPlayPatch : BytecodePatch(
-    listOf(
-        BackgroundPlaybackDisableFingerprint
-    )
+    listOf(BackgroundPlaybackDisableFingerprint)
 ) {
     override fun execute(context: BytecodeContext): PatchResult {
         BackgroundPlaybackDisableFingerprint.result!!.mutableMethod.addInstructions(