+++ src/main/kotlin/app/revanced/patches/reddit/customclients/relayforreddit/api/fingerprints/LoginActivityClientIdFingerprint.kt
@@ -1,0 +1,3 @@
+package app.revanced.patches.reddit.customclients.relayforreddit.api.fingerprints
+
+object LoginActivityClientIdFingerprint : AbstractClientIdFingerprint("&duration=permanent")