+++ src/main/kotlin/app/revanced/patches/reddit/customclients/syncforreddit/api/url/fingerprints/GetRedditImageUrlFingerprint.kt
@@ -1,0 +1,7 @@
+package app.revanced.patches.reddit.customclients.syncforreddit.api.url.fingerprints
+
+import app.revanced.patcher.fingerprint.method.impl.MethodFingerprint
+
+object GetRedditImageUrlFingerprint : MethodFingerprint(
+    strings = listOf("preview.redd.it")
+)