+++ src/main/kotlin/app/revanced/patches/shared/fingerprints/WatchWhileActivityFingerprint.kt
@@ -1,10 +1,13 @@
 package app.revanced.patches.shared.fingerprints
 
-
+import app.revanced.patcher.extensions.or
 import app.revanced.patcher.fingerprint.method.impl.MethodFingerprint
+import org.jf.dexlib2.AccessFlags
 
 object WatchWhileActivityFingerprint : MethodFingerprint(
+    accessFlags = AccessFlags.PUBLIC or AccessFlags.CONSTRUCTOR,
+    parameters = listOf(),
     customFingerprint = { methodDef, _ ->
-        methodDef.definingClass.endsWith("WatchWhileActivity;") && methodDef.name == "<init>"
+        methodDef.definingClass.endsWith("WatchWhileActivity;")
     }
 )