+++ src/main/kotlin/app/revanced/patches/syncforreddit/api/url/fingerprints/GetRedditUrlFingerprint.kt
@@ -1,7 +1,0 @@
-package app.revanced.patches.syncforreddit.api.url.fingerprints
-
-import app.revanced.patcher.fingerprint.method.impl.MethodFingerprint
-
-object GetRedditUrlFingerprint : MethodFingerprint(
-    strings = listOf("reddit.com")
-)