+++ src/main/kotlin/app/revanced/patches/youtube/interaction/seekbar/fingerprints/AccessibilityPlayerProgressTimeFingerprint.kt
@@ -1,23 +1,12 @@
 package app.revanced.patches.youtube.interaction.seekbar.fingerprints
 
 import app.revanced.patcher.extensions.or
-import app.revanced.patcher.fingerprint.method.impl.MethodFingerprint
 import app.revanced.patches.youtube.interaction.seekbar.patch.EnableSeekbarTappingResourcePatch
+import app.revanced.util.patch.LiteralValueFingerprint
 import org.jf.dexlib2.AccessFlags
-import org.jf.dexlib2.Opcode
-import org.jf.dexlib2.iface.instruction.WideLiteralInstruction
 
-
-object AccessibilityPlayerProgressTimeFingerprint : MethodFingerprint(
+object AccessibilityPlayerProgressTimeFingerprint : LiteralValueFingerprint(
     returnType = "L",
     accessFlags = AccessFlags.PUBLIC or AccessFlags.FINAL,
-    customFingerprint = { methodDef, _ ->
-        methodDef.implementation?.instructions?.any { instruction ->
-            if (instruction.opcode != Opcode.CONST) return@any false
-
-            val wideLiteral = (instruction as WideLiteralInstruction).wideLiteral
-
-            EnableSeekbarTappingResourcePatch.accessibilityPlayerProgressTime == wideLiteral
-        } ?: false
-    }
+    literal = EnableSeekbarTappingResourcePatch.accessibilityPlayerProgressTime
 )