+++ src/main/kotlin/app/revanced/patches/youtube/layout/hide/loadmorebutton/bytecode/fingerprints/HideLoadMoreButtonFingerprint.kt
@@ -1,13 +1,12 @@
 package app.revanced.patches.youtube.layout.hide.loadmorebutton.bytecode.fingerprints
 
 import app.revanced.patcher.extensions.or
-import app.revanced.patcher.fingerprint.method.impl.MethodFingerprint
+import app.revanced.util.patch.LiteralValueFingerprint
 import app.revanced.patches.youtube.layout.hide.loadmorebutton.resource.patch.HideLoadMoreButtonResourcePatch
 import org.jf.dexlib2.AccessFlags
 import org.jf.dexlib2.Opcode
-import org.jf.dexlib2.iface.instruction.WideLiteralInstruction
 
-object HideLoadMoreButtonFingerprint : MethodFingerprint(
+object HideLoadMoreButtonFingerprint : LiteralValueFingerprint(
     returnType = "V",
     accessFlags = AccessFlags.PUBLIC or AccessFlags.CONSTRUCTOR,
     parameters = listOf("L", "L", "L", "L"),
@@ -17,13 +16,5 @@ object HideLoadMoreButtonFingerprint : MethodFingerprint(
         Opcode.INVOKE_STATIC,
         Opcode.MOVE_RESULT_OBJECT
     ),
-    customFingerprint = { methodDef, _ ->
-        methodDef.implementation?.instructions?.any {
-            if (it.opcode != Opcode.CONST) return@any false
-
-            val literal = (it as WideLiteralInstruction).wideLiteral
-
-            literal == HideLoadMoreButtonResourcePatch.expandButtonDownId
-        } ?: false
-    }
+    literal = HideLoadMoreButtonResourcePatch.expandButtonDownId
 )