+++ src/main/kotlin/app/revanced/patches/youtube/layout/tabletminiplayer/fingerprints/MiniPlayerResponseModelSizeCheckFingerprint.kt
@@ -8,7 +8,7 @@ import org.jf.dexlib2.Opcode
 object MiniPlayerResponseModelSizeCheckFingerprint : MethodFingerprint(
     "L",
     AccessFlags.PUBLIC or AccessFlags.FINAL,
-    listOf("L", "L"),
+    listOf("Ljava/lang/Object;", "Ljava/lang/Object;"),
     listOf(
         Opcode.RETURN_OBJECT,
         Opcode.CHECK_CAST,